+++ docker/compose.sh
@@ -1,0 +1,6 @@
+#!/usr/bin/env bash
+SCRIPT_ROOT=$(cd $(dirname $0); pwd)
+
+cd $SCRIPT_ROOT
+
+docker compose -p lwdfx $@