+++ docker/logs.sh
@@ -1,0 +1,6 @@
+#!/usr/bin/env bash
+SCRIPT_ROOT=$(cd $(dirname $0); pwd)
+
+cd $SCRIPT_ROOT
+
+./compose.sh logs --tail 50 -f